chore: Refactor star rating component and add checkbox component
