+++ src/components/checkbox.tsx
@@ -1,0 +1,5 @@
+function Checkbox({ checked, onCheckedChange }: { checked: boolean; onCheckedChange: () => void }) {
+  return <input type='checkbox' checked={checked} onChange={onCheckedChange} />;
+}
+
+export default Checkbox;
